refactor: 로또 게임 작동 구조 개선 완료
